+++ hibernate-core/src/test/java/org/hibernate/test/cache/LockModeTest.java
@@ -49,33 +49,30 @@ public class LockModeTest extends BaseCoreFunctionalTestCase {
 	}
 
 	@Override
-	protected void prepareTest() throws Exception {
-		Session s = openSession();
-		s.beginTransaction();
+	protected void prepareTest() {
+		inTransaction(
+				s -> {
+					Company company1 = new Company( 1 );
+					s.save( company1 );
 
-		Company company1 = new Company( 1 );
-		s.save( company1 );
+					User user = new User( 1, company1 );
+					s.save( user );
 
-		User user = new User( 1, company1 );
-		s.save( user );
-
-		Company company2 = new Company( 2 );
-		s.save( company2 );
-
-		s.getTransaction().commit();
-		s.close();
+					Company company2 = new Company( 2 );
+					s.save( company2 );
+				}
+		);
 	}
 
 	@Override
-	protected void cleanupTest() throws Exception {
-		Session s = openSession();
-		s.beginTransaction();
+	protected void cleanupTest() {
+		inTransaction(
+				s -> {
+					s.createQuery( "delete from org.hibernate.test.cache.User" ).executeUpdate();
+					s.createQuery( "delete from org.hibernate.test.cache.Company" ).executeUpdate();
 
-		s.createQuery( "delete from org.hibernate.test.cache.User" ).executeUpdate();
-		s.createQuery( "delete from org.hibernate.test.cache.Company" ).executeUpdate();
-
-		s.getTransaction().commit();
-		s.close();
+				}
+		);
 	}
 
 	/**
